Remove `readable-stream`
This package was only being used in tests. The tests are always run in
`Node.js`, so I switched them to use the built-in Node.js stream API
instead.

I'm guessing this package was originally included to allow using this
package in either the browser or Node.js. But it wasn't setup that way.
I did consider using `readable-stream` in place of `stream` to make
browser compatibility easier for consumers, but that would have
required more effort to get working, as the version of
`readable-stream` in-use was outdated and incompatible with this
package. Plus it's fairly common for bundlers to replace built-in
modules like `stream` on-the-fly so it would have been of marginal
benefit.