Allow falsy initial state (#74)
This library has used an empty object as the default initial state ever
since the initial version, but in older versions that was only set if
the initial state was undefined. An initial state of `null` or any
other falsy value other than `undefined`  was still accepted.

In #38 (published as part of v6.0.2) this was accidentally changed.
Instead the empty object default was used for any falsy value passed in
as the initial state.

The pre-v6.0.2 behavior has now been restored; falsy initial state is
no longer overwritten.