Convert tests to TypeScript (#79)
We need to fix a bug with the types for ComposedStore, but before we can
do that, we need to verify that the fix works. Converting the tests to
TypeScript should help.

Note that doing this has revealed another bug with the type for the
ObservableStore constructor. We will address that bug in a future
commit.